Resolve the service class tests
